Add proxy as module
