+++ lib/pmcrypto.js
@@ -72,6 +72,8 @@ export { default as keyInfo, getSHA256Fingerprints } from './key/info';
 
 export { keyCheck } from './key/check';
 
+export { proxyTransform } from './key/proxy';
+
 export * from './constants';
 
 export const config = { debug: true };
